Change SSA JWKS endpoints and publish network cert
Our current understanding is:

The `org_jwks_uri` endpoint should provide the signing and transport
certificates for the organisation.

The `software_jwks_uri` endpoint should provide the signing and
transport certificates for the software.

Also changed some of the routes in order to better match those specified
in the Open Banking API spec:

https://github.com/OpenBankingUK/directory-api-specs/blob/master/directory-api-swagger.yaml

The specs also mention that the public key used to validate the SSA
signature should be retrieved from an endpoint that is not in the SSA
itself. However, there does not seem to be any mention of this in the
above Swagger spec.

It currently exists at `/directory/:ssa_id/key` until we can investigate
further.